+++ app/src/main/java/dk/shape/games/notifications/demo/dependency/MockNotificationSettingsDependencyProvider.kt
@@ -3,7 +3,9 @@ package dk.shape.games.notifications.demo.dependency
 import androidx.fragment.app.Fragment
 import dk.shape.games.notifications.demo.mock.*
 import dk.shape.games.notifications.presentation.NotificationSettingsConfig
+import dk.shape.games.notifications.presentation.NotificationSettingsSubjectConfig
 import dk.shape.games.notifications.presentation.NotificationSettingsSubjectFragment
+import dk.shape.games.notifications.presentation.SubjectNotificationStateData
 import dk.shape.games.toolbox_library.configinjection.ConfigProvider
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlin.time.ExperimentalTime
@@ -11,39 +13,61 @@ import kotlin.time.ExperimentalTime
 @ExperimentalCoroutinesApi
 @ExperimentalTime
 class MockNotificationSettingsDependencyProvider : ConfigProvider<NotificationSettingsConfig> {
+    override fun config(fragment: Fragment): NotificationSettingsConfig =
+        mockNotificationSettingsConfig
+}
+
+var savedNotificationListener: (SubjectNotificationStateData) -> Unit = {}
+
+@ExperimentalCoroutinesApi
+@ExperimentalTime
+private val mockNotificationSettingsConfig = NotificationSettingsConfig(
+    legacyNotificationsComponent = mockLegacyNotificationsComponent,
+    subjectNotificationsDataSource = mockSubjectNotificationsDataSource,
+    provideEventIdsForUserBetsAsync = {},
+    provideEventIdsForBetSlip = { null },
+    provideSubjectInfo = {
+        mockSubjectInfos
+    },
+    provideAppConfig = {
+        mockAppConfig
+    },
+    provideEvents = {
+        mockEvents
+    },
+    provideDeviceId = {
+        "device:1234"
+    },
+    onBackPressed = {},
+    onSubjectNotificationTypesClicked = { fragment, subjectNotificationTypesAction, notificationListener ->
+        NotificationSettingsSubjectFragment().apply {
+            arguments = NotificationSettingsSubjectFragment.Args.create(
+                subjectNotificationTypesAction,
+                MockNotificationSettingsSubjectDependencyProvider::class.java
+            )
+            show(
+                fragment.childFragmentManager,
+                NotificationSettingsSubjectFragment::class.java.simpleName
+            )
+        }
+        savedNotificationListener = notificationListener
+    },
+    onEventNotificationTypesClicked = { fragment, legacyEventNotificationTypesAction -> }
+)
+
+@ExperimentalTime
+class MockNotificationSettingsSubjectDependencyProvider :
+    ConfigProvider<NotificationSettingsSubjectConfig> {
     @ExperimentalCoroutinesApi
-    override fun config(fragment: Fragment): NotificationSettingsConfig {
-        return NotificationSettingsConfig(
-            legacyNotificationsComponent = mockLegacyNotificationsComponent,
-            subjectNotificationsDataSource = mockSubjectNotificationsDataSource,
-            provideEventIdsForUserBetsAsync = {},
-            provideEventIdsForBetSlip = { null },
-            provideSubjectInfo = {
-                mockSubjectInfos
-            },
-            provideAppConfig = {
-                mockAppConfig
-            },
-            provideEvents = {
-                mockEvents
-            },
+    override fun config(fragment: Fragment): NotificationSettingsSubjectConfig {
+        return NotificationSettingsSubjectConfig(
+            notificationsDataSource = mockSubjectNotificationsDataSource,
             provideDeviceId = {
                 "device:1234"
             },
-            onBackPressed = {},
-            onSubjectNotificationTypesClicked = { fragment, subjectNotificationTypesAction ->
-                NotificationSettingsSubjectFragment().apply {
-                    arguments = NotificationSettingsSubjectFragment.Args.create(
-                        mockNotificationSettingsSubjectAction,
-                        MockNotificationSettingsSubjectDependencyProvider::class.java
-                    )
-                    show(
-                        fragment.childFragmentManager,
-                        NotificationSettingsSubjectFragment::class.java.simpleName
-                    )
-                }
-            },
-            onEventNotificationTypesClicked = { fragment, legacyEventNotificationTypesAction -> }
+            onNotificationTypesChanged = { stateData ->
+                savedNotificationListener(stateData)
+            }
         )
     }
 }
@@ -71,7 +95,7 @@ class MockNotificationSettingsEmptyDependencyProvider : ConfigProvider<Notificat
                 "device:1234"
             },
             onBackPressed = {},
-            onSubjectNotificationTypesClicked = { fragment, subjectNotificationTypesAction -> },
+            onSubjectNotificationTypesClicked = { fragment, subjectNotificationTypesAction, listener -> },
             onEventNotificationTypesClicked = { fragment, legacyEventNotificationTypesAction -> }
         )
     }
