+++ notifications/src/main/java/dk/shape/games/notifications/presentation/NotificationSettingsSubjectConfig.kt
@@ -1,13 +1,14 @@
 package dk.shape.games.notifications.presentation
 
-import dk.shape.games.notifications.aliases.SubjectNotificationGroup
 import dk.shape.games.notifications.repositories.SubjectNotificationsDataSource
 
 // TODO: add description once API is stable
 
 data class NotificationSettingsSubjectConfig(
 
+    val notificationsDataSource: SubjectNotificationsDataSource,
+
     val provideDeviceId: suspend () -> String,
 
-    val notificationsDataSource: SubjectNotificationsDataSource
+    val onNotificationTypesChanged: (SubjectNotificationStateData) -> Unit
 )
