+++ notifications/src/main/java/dk/shape/games/notifications/presentation/NotificationSettingsSubjectFragment.kt
@@ -53,7 +53,10 @@ class NotificationSettingsSubjectFragment : ExpandableBottomSheetDialogFragment(
                                 subjectId = stateData.subjectId,
                                 subjectType = stateData.subjectType,
                                 notificationTypeIds = stateData.notificationTypeIdentifiers.toSet(),
-                                onSuccess = onSuccess,
+                                onSuccess = {
+                                    onSuccess()
+                                    config.onNotificationTypesChanged(stateData)
+                                },
                                 onError = onFailure
                             )
                         }
